% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_interval.R
\name{arg.atanx}
\alias{arg.atanx}
\title{Argument to atan(h) functions}
\usage{
arg.atanx(kd, ki = kd, d = 1)
}
\arguments{
\item{kd}{value at focal pole of decreasing items}

\item{ki}{value at focal pole of increasing items}

\item{d}{length of interval}
}
\description{
Argument to atan(h) functions
}
