% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{bounds.in.grid}
\alias{bounds.in.grid}
\title{Find bracketing points within a grid}
\usage{
bounds.in.grid(x0, xgrid)
}
\arguments{
\item{x0}{focal point}

\item{xgrid}{grid to search through}
}
\value{
list with \code{l}eft and \code{r}ight endpoints (or NA for open intervals)
}
\description{
Given a focal point, finds the positions that bracket it within a grid,
possibly equal at left endpoint (i.e. closer to -Inf)
}
