% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{bridge.mean}
\alias{bridge.mean}
\title{Bridge interval mean}
\usage{
bridge.mean(x0, xl, xr, vl, vr)
}
\arguments{
\item{x0}{focal point}

\item{xl}{left endpoint of interval}

\item{xr}{right endpoint of interval}

\item{vl}{value at left endpoint of interval}

\item{vr}{value at right endpoint of interval}
}
\description{
Bridge interval mean
}
