% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{bridge.sd}
\alias{bridge.sd}
\title{Bridge interval standard deviation}
\usage{
bridge.sd(x0, xl, xr)
}
\arguments{
\item{x0}{focal point}

\item{xl}{left endpoint of interval}

\item{xr}{right endpoint of interval}
}
\description{
Bridge interval standard deviation
}
