% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{Eumsi.optim.vec}
\alias{Eumsi.optim.vec}
\title{Optimization target for expected utility (inverse utility) of market share}
\usage{
Eumsi.optim.vec(x0s, gh.rule = fastGHQuad::gaussHermiteData(30), ...)
}
\description{
Optimization target for expected utility (inverse utility) of market share
}
