% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{Eumsi.optimize}
\alias{Eumsi.optimize}
\title{Find optimal entry location per expected inverse utility}
\usage{
Eumsi.optimize(
  vgrid,
  xgrid,
  term.last.optim = c(-4, 4),
  gh.rule = fastGHQuad::gaussHermiteData(30),
  ...
)
}
\arguments{
\item{vgrid}{values of other competitors}

\item{xgrid}{positions of other competitors}

\item{term.last.optim}{locations of last optima at terminals}

\item{gh.rule}{Gauss-Hermite rule to use across intergrations}

\item{...}{arguments passed to further functions}
}
\description{
Find optimal entry location per expected inverse utility
}
