% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{Eumsi.optimize.terminal}
\alias{Eumsi.optimize.terminal}
\title{Find optimal entry location on terminal intervals}
\usage{
Eumsi.optimize.terminal(x, sgn = 1, last.optim = 4, ...)
}
\arguments{
\item{x}{location of terminus}

\item{sgn}{search in positive or negative direction?}

\item{last.optim}{best guess is last optimum... (4 seems to be good baseline for a 0,0 start)}

\item{...}{arguments passed to further functions}
}
\description{
Find optimal entry location on terminal intervals
}
