% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_total.R
\name{ms.total}
\alias{ms.total}
\title{Total market share across the number line}
\usage{
ms.total(v0s, x0, vgrid, xgrid)
}
\arguments{
\item{v0s}{value of focal producer at their location (vectorized)}

\item{x0}{location of focal producer}

\item{vgrid}{values of other producers at their location}

\item{xgrid}{locations of other producers}
}
\description{
Total market share across the number line
}
