% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_total.R
\name{ms.total.trunc}
\alias{ms.total.trunc}
\title{Total market share across the number line (truncated)}
\usage{
ms.total.trunc(v0s, x0, vgrid, xgrid, trunc = 0.01)
}
\arguments{
\item{v0s}{value of focal producer at their location (vectorized)}

\item{x0}{location of focal producer}

\item{vgrid}{values of other producers at their location}

\item{xgrid}{locations of other producers}
}
\description{
Total market share across the number line (truncated)
}
