% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_interval.R
\name{msik.baseline.mpfr.wrap}
\alias{msik.baseline.mpfr.wrap}
\title{Market Share - Interval; k (exponentiated) formulation; naive mpfr-wrapped}
\usage{
msik.baseline.mpfr.wrap(kd, ki = kd, d = 1, prec = 1024)
}
\arguments{
\item{kd}{value at focal pole of decreasing items}

\item{ki}{value at focal pole of increasing items}

\item{d}{length of interval}
}
\description{
Market Share - Interval; k (exponentiated) formulation; naive mpfr-wrapped
}
