% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_interval.R
\name{msik.baseline}
\alias{msik.baseline}
\title{Market Share - Interval; k (exponentiated) formulation}
\usage{
msik.baseline(kd, ki = kd, d = 1)
}
\arguments{
\item{kd}{value at focal pole of decreasing items}

\item{ki}{value at focal pole of increasing items}

\item{d}{length of interval}
}
\description{
Market Share - Interval; k (exponentiated) formulation
}
