% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_interval.R
\name{msik.full.mpfr}
\alias{msik.full.mpfr}
\title{Market Share - Interval; k (exponentiated) formulation; fully inside mpfr land}
\usage{
msik.full.mpfr(kd, ki = kd, d = 1, prec = 1024)
}
\arguments{
\item{kd}{value at focal pole of decreasing items}

\item{ki}{value at focal pole of increasing items}

\item{d}{length of interval}
}
\description{
Market Share - Interval; k (exponentiated) formulation; fully inside mpfr land
}
