% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_interval.R
\name{msik}
\alias{msik}
\title{Market Share - Interval; k (exponentiated) formulation}
\usage{
msik(kd, ki, d, prec = 1024)
}
\arguments{
\item{kd}{value at focal pole of decreasing items}

\item{ki}{value at focal pole of increasing items}

\item{d}{length of interval}
}
\description{
Market Share - Interval; k (exponentiated) formulation
}
