% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_interval.R
\name{msiv}
\alias{msiv}
\title{Market Share - Interval; v (base-quality) formulation}
\usage{
msiv(v0s, vds, vis, d)
}
\arguments{
\item{v0s}{value of focal producer at near endpoint (vectorized)}

\item{vds}{value of increasing producers at near endpoint (producers with value decreasing in distance)}

\item{vis}{value of increasing producers at near endpoint (producers with value increasing in distance)}

\item{d}{length of interval}
}
\description{
By construction, v0 (company of interest) is always in the decreasing set
}
