% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_terminal.R
\name{mstk}
\alias{mstk}
\title{Market Share, Terminal, k (exponentiated) formula}
\usage{
mstk(k)
}
\arguments{
\item{k}{sum of exponentiated market values}
}
\description{
Market Share, Terminal, k (exponentiated) formula
}
