% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_terminal.R
\name{mstv}
\alias{mstv}
\title{Market Share, Terminal, v (value) formula}
\usage{
mstv(v0s, vs)
}
\arguments{
\item{v0s}{value of focal producer at terminal pole (vectorized)}

\item{vs}{value of other producers at terminal pole}
}
\description{
Market Share, Terminal, v (value) formula
}
