% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{open.sd}
\alias{open.sd}
\title{Open interval standard deviation, vectorized}
\usage{
\method{open}{sd}(x0, xl, xr)
}
\arguments{
\item{x0}{focal point}

\item{xl}{left endpoint of interval (NA for -Inf)}

\item{xr}{right endpoint of interval (NA for Inf)}
}
\description{
Open interval standard deviation, vectorized
}
