% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{relative.share}
\alias{relative.share}
\title{Relative share of output}
\usage{
relative.share(v0s, vs)
}
\arguments{
\item{v0s}{value of focal producer at this pole (vectorized)}

\item{vs}{values of other same-direction competition producers from this pole}
}
\description{
Relative share of output
}
