% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{u.inverse}
\alias{u.inverse}
\title{Inverse utility function for softmax market share}
\usage{
u.inverse(ms, a = 0.9)
}
\arguments{
\item{ms}{market share}

\item{a}{risk aversion adjustment (a < 1 -> more risk aversion)}
}
\description{
Inverts two-sided terminal market share,
adjusting for option value of fleeing towards infinity.
}
