% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{vectorize.args}
\alias{vectorize.args}
\title{Convert arguments to vectors of equal length}
\usage{
vectorize.args(...)
}
\arguments{
\item{...}{ensures that all arguments have same length, expanding vectors of length 1}
}
\description{
Convert arguments to vectors of equal length
}
