% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{wiener.moments}
\alias{wiener.moments}
\title{Wiener process first moments (mean, sd)}
\usage{
wiener.moments(x0, xgrid, vgrid)
}
\arguments{
\item{x0}{focal point}

\item{xgrid}{positions}

\item{vgrid}{values of positions}
}
\value{
list: mean, sd
}
\description{
Wiener process first moments (mean, sd)
}
